rename package from android
