+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.beercost",
+    package = "com.xdigu.beerprice",
 )
 
 android_resource(
     name = "res",
-    package = "com.beercost",
+    package = "com.xdigu.beerprice",
     res = "src/main/res",
 )
