+++ android/app/build.gradle
@@ -128,11 +128,11 @@ android {
     }
 
     defaultConfig {
-        applicationId "com.beercost"
+        applicationId "com.xdigu.beerprice"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
-        versionCode 1
-        versionName "1.0"
+        versionCode 101
+        versionName "1.0.1"
     }
     splits {
         abi {
