+++ android/app/src/main/java/com/xdigu/beerprice/MainApplication.java
@@ -1,4 +1,4 @@
-package com.beercost;
+package com.xdigu.beerprice;
 
 import android.app.Application;
 import android.content.Context;
@@ -7,6 +7,7 @@ import com.facebook.react.ReactApplication;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.soloader.SoLoader;
+import com.facebook.react.BuildConfig;
 import java.lang.reflect.InvocationTargetException;
 import java.util.List;
